Initial commit
SPI driver that supports both the microcontroller (uc) and the
switch (switch) SPI endpoints of the SJA1110.

During probing, the driver tries to initialize both the switch
(by uploading a static configuration file) and the uc (by uploading
a firmware binary). Default filenames are used, which are configurable
via Module parameters `uc_fw_upload` and `switch_cfg_upload`.

A sysfs interface exposes the ability to upload arbitrary static
configurations and firmwares to userspace.