Add option to reset the SJA1110 via RESET_SOFT_N
In case the RESET_SOFT_N pin is connected to some GPIO,
the SJA1110 can be reset by pulling low this pin.
This is an alternative to the previous method, which
used the ResetCtrl register of the SJA1110 for reset.

The GPIO needs to be configured in the device tree
via a new property called 'reset-gpio'. Register method
is used as a fallback.

SJA1110 reset is exposed via sysfs for manual resets.